apps/server/requirements.txt
ruff
fastapi
uvicorn
python-dotenv
tortoise-orm
tomlkit
aerich
PyJWT
passlib
bcrypt
pydantic[email]
unstructured[all-docs]
requests
